Yet more cleanup.
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@843 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/XtGlue.c
@@ -576,19 +576,19 @@ select_setup(int *nfds, fd_set *readfds, fd_set *writefds,
 	for (ip = inputs; ip != (input_t *)NULL; ip = ip->next) {
 		if ((unsigned long)ip->condition & InputReadMask) {
 			FD_SET(ip->source, readfds);
-			if (ip->source >= *nfds)
+			if ((int)ip->source >= *nfds)
 				*nfds = ip->source + 1;
 			r = 1;
 		}
 		if ((unsigned long)ip->condition & InputWriteMask) {
 			FD_SET(ip->source, writefds);
-			if (ip->source >= *nfds)
+			if ((int)ip->source >= *nfds)
 				*nfds = ip->source + 1;
 			r = 1;
 		}
 		if ((unsigned long)ip->condition & InputExceptMask) {
 			FD_SET(ip->source, exceptfds);
-			if (ip->source >= *nfds)
+			if ((int)ip->source >= *nfds)
 				*nfds = ip->source + 1;
 			r = 1;
 		}
